Convert decorators to the new conventions
This will enable compatibility with Zeitwerk and
latest versions of Rails.
