--- app/concerns/solidus_prototypes/permitted_attributes_concern.rb
@@ -1,17 +1,0 @@
-# frozen_string_literal: true
-
-module SolidusPrototypes
-  module PermittedAttributesConcern
-    extend ActiveSupport::Concern
-
-    included do
-      product_attributes << :prototype_id
-    end
-
-    module ClassMethods
-      def product_attributes
-        @@product_attributes
-      end
-    end
-  end
-end
