--- app/concerns/solidus_prototypes/prototype_product_display_permission_concern.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module SolidusPrototypes
-  module PrototypeProductDisplayPermissionConcern
-    def activate!
-      super
-      can [:display, :admin], Spree::Prototype
-    end
-  end
-end
