--- app/concerns/solidus_prototypes/prototype_product_management_permission_concern.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-module SolidusPrototypes
-  module PrototypeProductManagementPermissionConcern
-    def activate!
-      can :manage, Spree::Prototype
-      super
-    end
-  end
-end
