+++ app/decorators/models/solidus_prototypes/spree/option_type_decorator.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module SolidusPrototypes
+  module Spree
+    module OptionTypeDecorator
+
+      def self.prepended(base)
+        base.class_eval do
+          has_many :option_type_prototypes
+          has_many :prototypes, through: :option_type_prototypes
+        end
+      end
+
+      ::Spree::OptionType.prepend self
+    end
+  end
+end
