+++ app/decorators/models/solidus_prototypes/spree/permission_sets/product_display_decorator.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module SolidusPrototypes
+  module Spree
+    module PermissionSets
+      module ProductDisplayDecorator
+
+        def activate!
+          super
+          can [:display, :admin], ::Spree::Prototype
+        end
+
+        ::Spree::PermissionSets::ProductDisplay.prepend self
+      end
+    end
+  end
+end
+
+
