+++ app/decorators/models/solidus_prototypes/spree/permission_sets/product_management_decorator.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module SolidusPrototypes
+  module Spree
+    module PermissionSets
+      module ProductManagementDecorator
+
+        def activate!
+          can :manage, ::Spree::Prototype
+          super
+        end
+
+        ::Spree::PermissionSets::ProductManagement.prepend self
+      end
+    end
+  end
+end
