--- lib/decorators/spree/admin/products_decorator.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-Spree::Admin::ProductsController.include SolidusPrototypes::ProductsLoaderConcern
