--- lib/decorators/spree/option_type_decorator.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-Spree::OptionType.include SolidusPrototypes::OptionTypePrototypeConcern
