--- lib/decorators/spree/permission_sets_product_management_decorator.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-Spree::PermissionSets::ProductManagement.prepend SolidusPrototypes::PrototypeProductManagementPermissionConcern
