--- lib/decorators/spree/permitted_attributes_prototype_decorator.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-Spree::PermittedAttributes.include SolidusPrototypes::PermittedAttributesConcern
