--- lib/decorators/spree/products_decorator.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-Spree::Product.include SolidusPrototypes::ProductPrototypeConcern
