--- lib/decorators/spree/property_decorator.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-Spree::Property.include SolidusPrototypes::PropertyPrototypeConcern
