--- lib/decorators/spree/taxon_decorator.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-Spree::Taxon.include SolidusPrototypes::TaxonPrototypeConcern
