+++ lib/solidus_prototypes/engine.rb
@@ -4,25 +4,15 @@ require 'spree/core'
 
 module SolidusPrototypes
   class Engine < Rails::Engine
-    include SolidusSupport::EngineExtensions::Decorators
+    include SolidusSupport::EngineExtensions
 
     isolate_namespace ::Spree
 
     engine_name 'solidus_prototypes'
 
-    config.autoload_paths += %W(#{config.root}/lib/concerns)
-
     # use rspec for tests
     config.generators do |g|
       g.test_framework :rspec
     end
-
-    def self.activate
-      Dir.glob(File.join(File.dirname(__FILE__), '../../lib/decorators/**/*_decorator*.rb')) do |c|
-        Rails.configuration.cache_classes ? require(c) : load(c)
-      end
-    end
-
-    config.to_prepare(&method(:activate).to_proc)
   end
 end
